+++ src/components/post/Post.tsx
@@ -23,7 +23,7 @@ import ProductIcon from "../common_part/ProductIcon";
 import TagIcon from "../common_part/TagIcon";
 import MarkdownForm from "../common_part/MarkdownForm";
 import PostImage from "../common_part/PostImage";
-import { editProduct, fetchProduct, postProduct, reduceTagNum } from "../../firebase/firestore";
+import { editProduct, fetchProduct, postProduct, reduceTagNum } from "../../firebase-com/firestore";
 import { AuthContext } from "../../auth/AuthProvider";
 
 const storage = getStorage(app);
