+++ src/components/user/DisplayUserProductList.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { HStack, TabList, Tabs, Tab, TabPanels, TabPanel } from "@chakra-ui/react";
 
-import { fetchProductsUser, fetchUserInfos } from "../../firebase/firestore";
+import { fetchProductsUser, fetchUserInfos } from "../../firebase-com/firestore";
 import DisplayProductProps from "../common_part/DisplayProductProps";
 import DisplayProducts from "../common_part/DisplayProducts";
