+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
-    "compilerOptions": {
-    "target": "es2017",
+  "compilerOptions": {
+    "target": "es5",
     "lib": [
-        "dom",
-        "dom.iterable",
-        "esnext"
+      "dom",
+      "dom.iterable",
+      "esnext"
     ],
     "allowJs": true,
     "skipLibCheck": true,
@@ -19,8 +19,8 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react"
-    },
-    "include": [
-        "src"
-    ]
+  },
+  "include": [
+    "src"
+  ]
 }
